+++ R/se.R
@@ -261,10 +261,15 @@ se.svyglm.nb <- function(x, ...) {
 
 #' @rdname se
 #' @export
-se.icc.lme4 <- function(x, nsim = 100, ...) {
-  std_e_icc(x, nsim)
+se.sj_icc_merMod <- function(x, nsim = 100, ...) {
+  std_e_icc(x, nsim, adjusted = FALSE)
 }
 
+#' @rdname se
+#' @export
+se.sj_icc <- function(x, nsim = 100, ...) {
+  std_e_icc(x, nsim, adjusted = TRUE)
+}
 
 #' @export
 se.glmerMod <- function(x, ...) {
@@ -356,7 +361,7 @@ std_merMod <- function(fit) {
 }
 
 
-std_e_icc <- function(x, nsim) {
+std_e_icc <- function(x, nsim, adjusted = FALSE) {
   # check whether model is still in environment?
   obj.name <- attr(x, ".obj.name", exact = T)
   if (!exists(obj.name, envir = globalenv()))
@@ -376,11 +381,16 @@ std_e_icc <- function(x, nsim) {
   if (missing(nsim) || is.null(nsim)) nsim <- 100
 
   # get ICC, and compute bootstrapped SE, than return both
-  bstr <-
-    bootstr_icc_se(model_frame(fitted.model, fe.only = FALSE),
-                   nsim,
-                   model.formula,
-                   model.family)
+  bstr <- bootstr_icc_se(
+    model_frame(fitted.model, fe.only = FALSE),
+    nsim,
+    model.formula,
+    model.family,
+    adjusted
+  )
+
+  # get adjusted ICC only
+  if (adjusted) x <- x[[1]]
 
   # now compute SE and p-values for the bootstrapped ICC
   res <- data_frame(
@@ -390,7 +400,10 @@ std_e_icc <- function(x, nsim) {
     p.value = boot_p(bstr)[["p.value"]]
   )
 
-  structure(class = "sj_se_icc", list(result = res, bootstrap_data = bstr))
+  structure(
+    class = "sj_se_icc",
+    list(result = res, bootstrap_data = bstr, adjusted = adjusted)
+  )
 }
 
 
@@ -400,7 +413,7 @@ std_e_icc <- function(x, nsim) {
 #' @importFrom utils txtProgressBar
 #' @importFrom purrr map map_dbl
 #' @importFrom rlang .data
-bootstr_icc_se <- function(dd, nsim, formula, model.family) {
+bootstr_icc_se <- function(dd, nsim, formula, model.family, adjusted) {
   # create progress bar
   pb <- utils::txtProgressBar(min = 1, max = nsim, style = 3)
 
@@ -418,14 +431,20 @@ bootstr_icc_se <- function(dd, nsim, formula, model.family) {
           lme4::glmer(formula, data = x, family = model.family)
       }),
       # compute ICC(s) for each "bootstrapped" regression
-      icc = purrr::map(.data$models, icc)
+      icc = suppressMessages(purrr::map(.data$models, ~ icc(.x, adjusted = adjusted)))
     )
 
   # we may have more than one random term in the model, so we might have
   # multiple ICC values. In this case, we need to split the multiple icc-values
   # into multiple columns, i.e. one column per ICC value
-  icc_data <-
-    as.data.frame(matrix(unlist(purrr::map(dummy$icc, ~ .x)), nrow = nrow(dummy)))
+
+  if (adjusted) {
+    icc_data <-
+      as.data.frame(matrix(unlist(purrr::map(dummy$icc, ~ .x[[1]])), nrow = nrow(dummy)))
+  } else {
+    icc_data <-
+      as.data.frame(matrix(unlist(purrr::map(dummy$icc, ~ .x)), nrow = nrow(dummy)))
+  }
 
   # close progresss bar
   close(pb)